backend/requirements.txt
fastapi
uvicorn
tidalapi
pyyaml
python-dotenv
tqdm
sqlalchemy